backend/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy==1.24.3
opencv-python==4.8.1.78
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
ultralytics==8.0.196
torch==2.0.1+cu118
torchvision==0.15.2+cu118
alembic==1.13.1 

